* commit 'acb1340527c3422d1341c168bb39909ecd83486d':
  fix cache params, fix text, fix ajax exceptions
  fix orders
  кеш, расширенные ошибки, логи, рефакторинг, чистка

+++ src/PersonalBundle/AjaxController/AddressController.php
@@ -6,12 +6,26 @@
 
 namespace FourPaws\PersonalBundle\AjaxController;
 
-use FourPaws\App\Response\JsonErrorResponse;
+use Adv\Bitrixtools\Tools\Log\LoggerFactory;
+use Bitrix\Main\ObjectPropertyException;
+use Bitrix\Main\Security\SecurityException;
+use FourPaws\App\Exceptions\ApplicationCreateException;
 use FourPaws\App\Response\JsonResponse;
 use FourPaws\App\Response\JsonSuccessResponse;
+use FourPaws\AppBundle\Exception\EmptyEntityClass;
+use FourPaws\AppBundle\Exception\NotFoundException;
+use FourPaws\AppBundle\Service\AjaxMess;
 use FourPaws\PersonalBundle\Service\AddressService;
+use FourPaws\UserBundle\Exception\BitrixRuntimeException;
+use FourPaws\UserBundle\Exception\ConstraintDefinitionException;
+use FourPaws\UserBundle\Exception\InvalidIdentifierException;
+use FourPaws\UserBundle\Exception\NotAuthorizedException;
+use FourPaws\UserBundle\Exception\ValidationException;
+use FourPaws\UserBundle\Service\UserAuthorizationInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\DependencyInjection\Exception\ServiceCircularReferenceException;
+use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -26,28 +40,39 @@ class AddressController extends Controller
      * @var AddressService
      */
     private $addressService;
-    
+
+    /** @var AjaxMess */
+    private $ajaxMess;
+    /** @var UserAuthorizationInterface */
+    private $userAuthorization;
+
     public function __construct(
-        AddressService $addressService
+        AddressService $addressService,
+        UserAuthorizationInterface $userAuthorization,
+        AjaxMess $ajaxMess
     ) {
         $this->addressService = $addressService;
+        $this->userAuthorization = $userAuthorization;
+        $this->ajaxMess = $ajaxMess;
     }
-    
+
     /**
      * @Route("/add/", methods={"POST"})
      * @param Request $request
      *
      * @return JsonResponse
+     * @throws \RuntimeException
      */
-    public function addAction(Request $request) : JsonResponse
+    public function addAction(Request $request): JsonResponse
     {
+        if (!$this->userAuthorization->isAuthorized()) {
+            return $this->ajaxMess->getNeedAuthError();
+        }
         $data = $request->request->all();
         if (empty($data)) {
-            return JsonErrorResponse::createWithData(
-                'Не указаны данные для добавления',
-                ['errors' => ['emptyData' => 'Не указаны данные для добавления']]
-            );
+            return $this->ajaxMess->getEmptyDataError();
         }
+
         try {
             if ($this->addressService->addFromArray($data)) {
                 return JsonSuccessResponse::create(
@@ -57,39 +82,41 @@ class AddressController extends Controller
                     ['reload' => true]
                 );
             }
-        } catch (\Exception $e) {
-            return JsonErrorResponse::createWithData(
-                $e->getMessage(),
-                ['errors' => ['systemError' => $e->getMessage()]]
-            );
+        } catch (BitrixRuntimeException $e) {
+            return $this->ajaxMess->getAddError($e->getMessage());
+        } catch (EmptyEntityClass $e) {
+            return $this->ajaxMess->getAddError();
+        } catch (NotAuthorizedException $e) {
+            return $this->ajaxMess->getNeedAuthError();
+        } catch (ValidationException|InvalidIdentifierException|ConstraintDefinitionException $e) {
+            $logger = LoggerFactory::create('params');
+            $logger->error('Ошибка параметров - ' . $e->getMessage());
+        } catch (ApplicationCreateException|ServiceNotFoundException|ServiceCircularReferenceException|\RuntimeException|\Exception $e) {
+            $logger = LoggerFactory::create('system');
+            $logger->critical('Ошибка загрузки сервисов - ' . $e->getMessage());
         }
-        
-        return JsonErrorResponse::createWithData(
-            'Непредвиденная ошибка. Пожалуйста, обратитесь к администратору сайта',
-            ['errors' => ['systemError' => 'Непредвиденная ошибка. Пожалуйста, обратитесь к администратору сайта']]
-        );
+
+
+        return $this->ajaxMess->getSystemError();
     }
-    
+
     /**
      * @Route("/update/", methods={"POST"})
      * @param Request $request
      *
      * @return JsonResponse
      */
-    public function updateAction(Request $request) : JsonResponse
+    public function updateAction(Request $request): JsonResponse
     {
+        if (!$this->userAuthorization->isAuthorized()) {
+            return $this->ajaxMess->getNeedAuthError();
+        }
         $data = $request->request->all();
         if (empty($data)) {
-            return JsonErrorResponse::createWithData(
-                'Не указаны данные для обновления',
-                ['errors' => ['emptyData' => 'Не указаны данные для обновления']]
-            );
+            return $this->ajaxMess->getEmptyDataError();
         }
         if ((int)$data['ID'] < 1) {
-            return JsonErrorResponse::createWithData(
-                'Не указан элемент для обновления',
-                ['errors' => ['emptyIdError' => 'Не указан элемент для обновления']]
-            );
+            return $this->ajaxMess->getNotIdError(' для обновления');
         }
         try {
             if ($this->addressService->update($data)) {
@@ -100,29 +127,40 @@ class AddressController extends Controller
                     ['reload' => true]
                 );
             }
-        } catch (\Exception $e) {
+        } catch (SecurityException|NotFoundException $e) {
+            return $this->ajaxMess->getSecurityError();
+            /** показываем системную ошибку */
+        } catch (BitrixRuntimeException $e) {
+            return $this->ajaxMess->getUpdateError($e->getMessage());
+        } catch (EmptyEntityClass $e) {
+            return $this->ajaxMess->getUpdateError();
+        } catch (NotAuthorizedException $e) {
+            return $this->ajaxMess->getNeedAuthError();
+        } catch (ValidationException|InvalidIdentifierException|ConstraintDefinitionException|ObjectPropertyException $e) {
+            $logger = LoggerFactory::create('params');
+            $logger->error('Ошибка параметров - ' . $e->getMessage());
+        } catch (ApplicationCreateException|ServiceNotFoundException|ServiceCircularReferenceException|\RuntimeException|\Exception $e) {
+            $logger = LoggerFactory::create('system');
+            $logger->critical('Ошибка загрузки сервисов - ' . $e->getMessage());
         }
-        
-        return JsonErrorResponse::createWithData(
-            'Непредвиденная ошибка. Пожалуйста, обратитесь к администратору сайта',
-            ['errors' => ['systemError' => 'Непредвиденная ошибка. Пожалуйста, обратитесь к администратору сайта']]
-        );
+
+        return $this->ajaxMess->getSystemError();
     }
-    
+
     /**
      * @Route("/delete/", methods={"GET"})
      * @param Request $request
      *
      * @return JsonResponse
      */
-    public function deleteAction(Request $request) : JsonResponse
+    public function deleteAction(Request $request): JsonResponse
     {
+        if (!$this->userAuthorization->isAuthorized()) {
+            return $this->ajaxMess->getNeedAuthError();
+        }
         $delId = (int)$request->get('id');
         if ($delId < 1) {
-            return JsonErrorResponse::createWithData(
-                'Не указан элемент для удаления',
-                ['errors' => ['emptyIdError' => 'Не указан элемент для удаления']]
-            );
+            return $this->ajaxMess->getNotIdError(' для удаления');
         }
         try {
             if ($this->addressService->delete($delId)) {
@@ -133,12 +171,20 @@ class AddressController extends Controller
                     ['reload' => true]
                 );
             }
-        } catch (\Exception $e) {
+        } catch (SecurityException|NotFoundException $e) {
+            return $this->ajaxMess->getSecurityError();
+        } catch (BitrixRuntimeException $e) {
+            return $this->ajaxMess->getDeleteError($e->getMessage());
+        } catch (NotAuthorizedException $e) {
+            return $this->ajaxMess->getNeedAuthError();
+        } catch (ValidationException|InvalidIdentifierException|ConstraintDefinitionException|ObjectPropertyException $e) {
+            $logger = LoggerFactory::create('params');
+            $logger->error('Ошибка параметров - ' . $e->getMessage());
+        } catch (ApplicationCreateException|ServiceNotFoundException|ServiceCircularReferenceException|\RuntimeException|\Exception $e) {
+            $logger = LoggerFactory::create('system');
+            $logger->critical('Ошибка загрузки сервисов - ' . $e->getMessage());
         }
-        
-        return JsonErrorResponse::createWithData(
-            'Непредвиденная ошибка. Пожалуйста, обратитесь к администратору сайта',
-            ['errors' => ['systemError' => 'Непредвиденная ошибка. Пожалуйста, обратитесь к администратору сайта']]
-        );
+
+        return $this->ajaxMess->getSystemError();
     }
 }
